UI overhaul: modern theme, responsive layout, animations, forms, and …

+++ EnvelopeDescription/models.py
@@ -8,8 +8,8 @@ class EnvelopeDescription(models.Model):
     username = models.ForeignKey(User, on_delete=models.CASCADE)  # linked to username indirectly
     EnvelopeName = models.ForeignKey('Envelopes.Envelope_Home', on_delete=models.CASCADE)  # Limits to current user
     Description = models.TextField()
-    Money_Spent = models.IntegerField()
-    Money_Remaining = models.IntegerField()
+    Money_Spent = models.DecimalField(max_digits=10, decimal_places=2)
+    Money_Remaining = models.DecimalField(max_digits=10, decimal_places=2)
     created_at = models.DateTimeField(auto_now_add=True)
 
     class Meta:
